% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_scale.R
\name{center_scale}
\alias{center_scale}
\title{Center scale from PLSISE}
\usage{
center_scale(x)
}
\arguments{
\item{x}{a dataframe of numeric to center and scale}
}
\value{
mean : means of columns
 sd : standard deviations of columns
 Xk : Centered and scaled X dataframe
}
\description{
Center and scale a numeric data.frame
}
