% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_report.R
\name{classification_report}
\alias{classification_report}
\title{Classification report from PLSISE}
\usage{
classification_report(y, ypred)
}
\arguments{
\item{y}{True values of Y}

\item{ypred}{Predicted values of Y}
}
\value{
The classification report
}
\description{
Show the classification report of predictions and true values of an Y variable
}
