% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validation}
\alias{cross_validation}
\title{Cross validation from PLSISE}
\usage{
cross_validation(formula, data, ncomp, cv = 5, method = "splitTrainTest")
}
\arguments{
\item{formula}{an object of class "formula", symbolic description of model 
to be fitted.}

\item{data}{a data frame containing the variables in model.}

\item{ncomp}{an integer: the number of components to keep in model.}

\item{cv}{an integer: the number of folds for the cross validation method}

\item{method}{Method used to split the dataset in folds, use 
"splitTrainTest" to create random train and test datasets at each validation,
use "kFold" to split "data" in "cv" folds and use one of them as test at 
each validation.}
}
\value{
MeanFscore : mean of fscore on validation,
 BestFscore : max fscore on validation,  
 BestModel : model with best fscore on validation
}
\description{
Run cross validation of the PLS-DA classifier on a dataset. Use the fit and
 predict function from PLSISE package. Compare PLSDA model by fscore
}
