% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dummies.R
\name{get_dummies}
\alias{get_dummies}
\title{Get dummies from PLSISE}
\usage{
get_dummies(x)
}
\arguments{
\item{x}{a factor containing values to make dummies of}
}
\value{
a dataframe containing the dummies
}
\description{
Create dummies of a factor variable
}
