% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pls_individuals}
\alias{pls_individuals}
\title{PLS individuals from PLSISE}
\usage{
pls_individuals(PLSDA, Axis_1 = 1, Axis_2 = 2)
}
\arguments{
\item{PLSDA}{an object of class PLSDA : a fitted with PLS-DA
method (fit) model.}

\item{Axis_1}{an integer : the number of the component on x axis}

\item{Axis_2}{an integer : the number of the component on y axis}
}
\value{
The individuals on factorial plan plot
}
\description{
Show individuals on factorial plan
}
