% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pls_variables}
\alias{pls_variables}
\title{PLS variables from PLSISE}
\usage{
pls_variables(PLSDA, Axis_1 = 1, Axis_2 = 2)
}
\arguments{
\item{PLSDA}{an object of class PLSDA : a fitted with PLS-DA
method (fit) model.}

\item{Axis_1}{an integer : the number of the component on x axis}

\item{Axis_2}{an integer : the number of the component on y axis}
}
\value{
The variables on factorial plan plot
}
\description{
Show variables on factorial plan
}
