% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.PLSDA}
\alias{predict.PLSDA}
\title{Fitting PLS-DA model}
\usage{
\method{predict}{PLSDA}(PLSDAmodel, newdata, type)
}
\arguments{
\item{PLSDAmodel}{an object of class PLSDA : a fitted with PLS-DA
method model.}

\item{newdata}{A data frame in which to look for variables with
which to predict.}

\item{type}{type of prediction : "class" for predicted class or
"posterior" for probabilities}
}
\description{
`predict()` Predicted values based on PLSDA object.
}
