% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot from PLSISE}
\usage{
scree_plot(PLSDA, method = "kaiser")
}
\arguments{
\item{PLSDA}{an object of class PLSDA : a fitted with PLS-DA
method (fit) model.}

\item{method}{The criterion for component selection, "kaiser" or 
"broken_sticks"}
}
\value{
The scree plot
}
\description{
Show scree plot of PLSDA : X object
}
