% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_variables.R
\name{select_variable}
\alias{select_variable}
\title{selection variables from PLSISE}
\usage{
select_variable(DF, cible, alpha = 0.05)
}
\arguments{
\item{DF}{A data frame in which to look for variables with
which to predict.}

\item{cible}{vector Y representing the different modalities for the membership classes}

\item{alpha}{significance level for the variables to be included in the selection}
}
\value{
lst_Var_Selected variables selected by the algorithm
}
\description{
Predict class, probabilities or values of X data based on a fitted PLS-DA model
}
