% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.PLSDA}
\alias{summary.PLSDA}
\title{summary.PLSDA from PLSISE}
\usage{
\method{summary}{PLSDA}(PLSDA)
}
\arguments{
\item{PLSDA}{a PLSDA object to print summary}
}
\description{
Summary of PLSDA object
}
