% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{x_plot}
\alias{x_plot}
\title{X plot from PLSISE}
\usage{
x_plot(data, varX, varY, class)
}
\arguments{
\item{data}{a data.frame object}

\item{varX}{Number or name of variable to represent on x axis}

\item{varY}{Number or name of variable to represent on y axis}

\item{class}{A factor : class of individuals}
}
\value{
The plot of individuals on variables plan
}
\description{
Show individuals by variables of a dataset
}
