% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fitting PLS-DA model}
\usage{
fit(formula, data, ncomp = 2)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be
coerced to that class): a symbolic description of the
model to be fitted.}

\item{data}{a data frame, (or object coercible
by as.data.frame to a data frame) containing the variables
in the model.}

\item{ncomp}{an integer: the number of components to keep in the model.}
}
\value{
an object of class PLSDA with the fitted model and it Attributes.
}
\description{
`fit()` Fit the PLS-DA Classification to data, it can be used to carry
 classification on a target variable with 2 or more modalities
}
