% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Fitting PLS-DA model}
\usage{
predict(PLSDA, newdata, type = "class")
}
\arguments{
\item{PLSDA}{an object of class PLSDA : a fitted with PLS-DA
method (fit) model.}

\item{newdata}{A data frame in which to look for variables with
which to predict.}

\item{type}{type of prediction : "class" for predicted class,
"posterior" for probabilities or "value" for values of predictions#'}
}
\value{
Y predictions for newdata as required in type parameter
}
\description{
`predict()` Predicted values based on PLSDA object.
}
